% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{data_process}
\alias{data_process}
\title{Title Data Processing for the model construction
Description This function processes the input data frame by checking the data types of its features and converting them to numeric if they meet certain conditions. It also counts the number of missing values in each row and column, and returns a list containing continuous features, discrete features, and NA counts.}
\usage{
data_process(Dat, pd = 5, na_threshold = 0.5)
}
\arguments{
\item{Dat}{A dataframe containing the variable to be processed.}

\item{pd}{A positive integer that is used to determine the minimum number of unique values required for a variable to be considered continuous. The default value is 5.}

\item{na_threshold}{A numeric value between 0 and 1 that represents the threshold for the proportion of missing values allowed in a variable. The default value is 0.5.}
}
\value{
A dataframe with the same structure as the input data frame, but with the data types of the specified features converted to numeric.
}
\description{
Title Data Processing for the model construction
Description This function processes the input data frame by checking the data types of its features and converting them to numeric if they meet certain conditions. It also counts the number of missing values in each row and column, and returns a list containing continuous features, discrete features, and NA counts.
}
\examples{

data(test_data)

# Process the test_data with default parameters
Dat <- data_process(test_data)

}
