% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Real-world human clinical data}
\format{
## `test_data`
A data frame with 150 rows and 17 columns:
\describe{
  \item{country}{China}
  ...
}
}
\usage{
test_data
}
\description{
Real-world human clinical data
}
\examples{
  head(test_data)
}
\keyword{datasets}
